Updated packages so animated gif now works again
I was getting errors in the generation of animated gif, similar to
what was reported here:

https://github.com/JuliaIO/FFMPEG.jl/issues/28

which eventually I solved by upgradsing all the packages in the Pluto
package.
This was done via the GUI by going to the 'Using Plots' chunk, and
selecting to update packages.  This could be seen to update several
packages and now code is working agin.

https://github.com/fonsp/Pluto.jl/wiki/%F0%9F%8E%81-Package-management#updating-packages